requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
pandas>=2.0.0
numpy>=1.24.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn